+++ avalon/urls.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.urls import path
 from django.views.generic import RedirectView
 
-from catalog.api import api
+from api.api import api
 
 urlpatterns = [
     path("", RedirectView.as_view(url="catalog/", permanent=True)),
